未修改的文件，保持它的encoding # Reviewed, transaction id: 41275

+++ packages/@bkui/cli-service/src/tools/rust/emit/index.ts
@@ -26,6 +26,10 @@ export const emit = async (context: IContext) => {
       });
     }
     // 写入文件
-    fs.writeFileSync(outputFilePath, file.content);
+    if (file.keepOriginFile) {
+      fs.cpSync(file.originRelativeFilePath, outputFilePath);
+    } else {
+      fs.writeFileSync(outputFilePath, file.content);
+    }
   }
 };
